chore: switch to amqprs crate + fix compilation (Rust 1.86 compat)
- Replaced lapin with amqprs (lighter dep tree, compatible with rustc 1.86)
- Added url crate for AMQP URL parsing
- Fixed AmqpConfig missing in devnet_default()
- Fixed usize→u64 cast for reveals/transfers count
- AMQP_URL env var override for SOPS secret injection
- Updated Cargo.lock (targeted deps only, not full update)

+++ components/ordinals/src/core/pipeline/processors/inscription_indexing.rs
@@ -241,8 +241,8 @@ pub async fn index_block(
                 &amqp.routing_key,
                 block_height,
                 &block.block_identifier.hash,
-                reveals_count,
-                transfers_count,
+                reveals_count as u64,
+                transfers_count as u64,
                 stopwatch.elapsed().as_millis() as u64,
             )
             .await
